feat: ✨ Add useGlobalLeaflet option prop
The useGlobalLeaflet option prop allows the user to mount the Lmap component by using the globally available leaflet instance instead of the one imported from the ESM file. This allows some plugins that require to have a global leaflet to be present to work with this component. By default the option is set to false.