Make `useGlobalLeaflet` prop of `LMap` only
Previously `useGlobalLeaflet` was an element of the `options` object on
every component, but only `LMap` actually did anything with the value.
Now the configuration option is explicitly only a prop on `LMap`, which
`provide`s its value to its children so that all components comprising a
single map will have the same setting value.